CDPPOC-9 rename deriveCaseState as deriveDisplayTextFromCaseState
