+++ microfrontends/case-admin-mfe/src/utils/deriveDisplayTextFromCaseState.js
